Add exceptions for items delete

+++ src/main/java/seedu/easylog/commands/itemscommands/ItemsDeleteCommand.java
@@ -2,16 +2,20 @@ package seedu.easylog.commands.itemscommands;
 
 import seedu.easylog.common.Constants;
 import seedu.easylog.exceptions.EmptyNumberException;
+import seedu.easylog.exceptions.InvalidNumberException;
 
 public class ItemsDeleteCommand extends ItemsCommand {
 
-    public void execute(String itemsArg) throws EmptyNumberException {
+    public void execute(String itemsArg) throws EmptyNumberException, InvalidNumberException {
         if (itemsArg.equals("")) {
             throw new EmptyNumberException();
         }
         int index = Integer.parseInt(itemsArg) - Constants.ARRAY_OFFSET;
-        ui.showDeletedItem(itemManager.getItem(index));
         int size = itemManager.getSize();
+        if ((index < 0) || (index >= size)) {
+            throw new InvalidNumberException();
+        }
+        ui.showDeletedItem(itemManager.getItem(index));
         itemManager.deleteItem(index);
         assert itemManager.getSize() == size - 1 : "After a valid deletion, one item is deleted.";
     }
